requirements.txt
# LangChain with OpenAI
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0

# Additional utilities
pydantic>=2.0.0
typing-extensions>=4.0.0

# Environment variables
python-dotenv

# Database
psycopg2-binary
# upstash-redis  # 保留用于数据迁移

# Web scraping
requests
beautifulsoup4

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0